requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
discord.py==1.5.1
google-auth==1.23.0
google-auth-oauthlib==0.4.2
gspread==3.6.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.6.0
python-dotenv==0.15.0
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1
